chore: refactor to make test code reusable across versions
